pci: fix spdk_pci_device_get_dev() accessor
spdk_pci_device_get_dev() needs to access the 'dev' member of a struct
pci_device, but the pointer to the struct is also called 'dev' in the
macro argument list.  Change it to something else so the 'dev' member
doesn't get replaced by the macro argument.

This is the minimal bug fix; more cleanup will happen later as the PCI
interface is refactored.

Change-Id: I37b4a303f980d40c5b1528293ecfd22ad0be9dc7
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>