test/bdevio: detach from build
With the I/O channel changes, this test needs to be rewritten
to be event-based.  bdevio uses the spdk_bdev_do_work() function
to poll for completions, which is built on the check_io functions
that are going away when we move to using I/O channels.

Do not delete the code from the tree - just detach it from the
build and the test scripts for now.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I88674988db6ccb3673faf7eb5b3e79b403059fa4